Fixed #2
The bug was non ops weren't able to use warps. For now, I've added the permission node:  essentials.warp.use. I plan on removing the essentials.warps.(warpname) feature in the near future due to this, but I do not know what I'll do with this feature for this stage. Some people don't realise how to give players permissions. Maybe I should put a soft dependency for PurePerms so they can do permissions as well as anything else, but this fix should be a lot easier now. So no need to worry that non ops are unable to warp to places. I will add a tutorial regarding this soon. If this does not fix #2, then I don't know what will, xD.