Dropped support for ALPHA10
This is because from Pocketmine build 516 - The latest have completely changed in these past moments. Which I had then to drop support for ALPHA10, and complete support for ALPHA11 only. You can however, use AllAPILoader plugin, allowing you to have plugins on a outdated server version with updated plugins, but please don't expect this to work with Older versions of Pocketmine. If it doesn't work, you won't be provided support with these issues. I'm currently supporting ALPHA11, as ALPHA10 is quite old, and can not be provided support considering we're constantly updating our plugin to the latest codes, API's, and so much more. Please bare with us though, we still plan on fixing certain codes, and what not. We usually update this plugin almost everyday, to get the updated plugin you need. This EssentialsPE plugin is the official and most updated of EssentialsPE you'll be able to ever find. I continued this project because the people who originally authored this plugin are very inactive, so I decided to continue this. I hope you all understand the reasons on why I continued this as a whole. Anyways, again, I only made support for ALPHA11. Possibly ALPHA12 once ALPHA12 arrives to pocketmine.